+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Allow using `add -c` for parachain chainSpecs of well-known relay-chains [#568](https://github.com/polkadot-api/polkadot-api/pull/568)
+- Use `@polkadot-api/polkadot-sdk-compat` with WS connections
 
 ### Breaking