+++ packages/json-rpc/json-rpc-provider-proxy/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Changed build to show individual files on source map explorers [#567](https://github.com/polkadot-api/polkadot-api/pull/567)
 
+### Breaking
+
+- Remove support for `transactionWatch`
+
 ## 0.1.0 - 2024-05-03
 
 ### Fixed