+++ packages/json-rpc/ws-provider/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Changed build to show individual files on source map explorers [#567](https://github.com/polkadot-api/polkadot-api/pull/567)
+- patch dependencies
 
 ## 0.1.0 - 2024-05-03